+++ tests/Fixtures/UpdatePostRequest.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Spawnflow\Tests\Fixtures;
+
+use Spawnflow\Http\SpawnflowFormRequest;
+
+class UpdatePostRequest extends SpawnflowFormRequest
+{
+    protected string $subject = 'posts';
+}